Move color scheme definitions to the same file
