+++ app/core/core-design-system/src/main/kotlin/com/hedvig/android/core/designsystem/material3/HedvigMaterial3Theme.kt
@@ -1,6 +1,7 @@
 package com.hedvig.android.core.designsystem.material3
 
 import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.material3.ColorScheme
 import androidx.compose.material3.LocalContentColor
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.darkColorScheme
@@ -8,6 +9,21 @@ import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.graphics.Color
+import com.hedvig.android.core.designsystem.newtheme.amber_100
+import com.hedvig.android.core.designsystem.newtheme.amber_400
+import com.hedvig.android.core.designsystem.newtheme.amber_600
+import com.hedvig.android.core.designsystem.newtheme.amber_800
+import com.hedvig.android.core.designsystem.newtheme.blue_100
+import com.hedvig.android.core.designsystem.newtheme.blue_400
+import com.hedvig.android.core.designsystem.newtheme.blue_600
+import com.hedvig.android.core.designsystem.newtheme.blue_800
+import com.hedvig.android.core.designsystem.newtheme.blue_900
+import com.hedvig.android.core.designsystem.newtheme.green_100
+import com.hedvig.android.core.designsystem.newtheme.green_400
+import com.hedvig.android.core.designsystem.newtheme.green_600
+import com.hedvig.android.core.designsystem.newtheme.green_800
+import com.hedvig.android.core.designsystem.newtheme.greyscale_0
+import com.hedvig.android.core.designsystem.newtheme.greyscale_1000
 import com.hedvig.android.core.designsystem.newtheme.hedvigTonalPalette
 import com.hedvig.android.core.designsystem.theme.lavender_300
 import com.hedvig.android.core.designsystem.theme.lavender_400
@@ -109,4 +125,49 @@ private val DarkColorScheme = darkColorScheme(
 //  outline = no equivalent,
   outlineVariant = hedvigTonalPalette.greyscale700,
   scrim = hedvigTonalPalette.greyscale0,
+)
+
+internal fun darkHedvigColorScheme(
+  colorScheme: ColorScheme,
+) = HedvigMaterial3ColorScheme(
+  containedButtonContainer = colorScheme.tertiary,
+  onContainedButtonContainer = colorScheme.onTertiary,
+
+  // In the comments are the light mode colors, showing the equivalent and how it was chosen
+  warningElement = amber_400, // amber_600,
+  onWarningElement = greyscale_0, // greyscale_1000,
+  warningContainer = amber_800, // amber_100,
+  onWarningContainer = amber_100, // amber_800,
+
+  typeElement = green_400, // green_600,
+  onTypeElement = greyscale_0, // greyscale_1000,
+  typeContainer = green_800, // green_100,
+  onTypeContainer = green_100, // green_800,
+
+  infoElement = blue_400, // blue_600,
+  onInfoElement = greyscale_1000, // greyscale_1000,
+  infoContainer = blue_800, // blue_100,
+  onInfoContainer = blue_100, // blue_900,
+)
+
+internal fun lightHedvigColorScheme(
+  colorScheme: ColorScheme,
+) = HedvigMaterial3ColorScheme(
+  containedButtonContainer = colorScheme.primary,
+  onContainedButtonContainer = colorScheme.onPrimary,
+
+  warningElement = amber_600,
+  onWarningElement = greyscale_1000,
+  warningContainer = amber_100,
+  onWarningContainer = amber_800,
+
+  typeElement = green_600,
+  onTypeElement = greyscale_1000,
+  typeContainer = green_100,
+  onTypeContainer = green_800,
+
+  infoElement = blue_600,
+  onInfoElement = greyscale_1000,
+  infoContainer = blue_100,
+  onInfoContainer = blue_900,
 )
